+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 * Initial release with minimally working StackStorm Vagrant box, created from Packer build pipeline
 * Add first system Vagrant-focused integration tests, tie them with the build (#5)
 * Add custom MOTD/welcome message after logging in to console (#15)
-* Add continuous integration (#19).
+* Add StackStorm infrastructure integration tests, ship with new `st2-integration-tests` executable available to user (#20)
+* Install Virtualbox guest additions
+* Add continuous integration (#19)